--- a/core/build.gradle
+++ b/core/build.gradle
@@ -6,7 +6,7 @@ dependencies {
     
     testImplementation libraries.test.jUnit
     testImplementation libraries.test.mockito
-    testImplementation libraries.test.fest
+    testImplementation libraries.test.truth
 }
 
 android {