--- a/core/src/test/java/com/novoda/merlin/ConnectCallbackManagerTest.java
+++ b/core/src/test/java/com/novoda/merlin/ConnectCallbackManagerTest.java
@@ -11,13 +11,12 @@ import static org.mockito.Mockito.verify;
 
 public class ConnectCallbackManagerTest {
 
-    private Register<Connectable> connectables;
+    private final Register<Connectable> connectables = new Register<>();
 
     private ConnectCallbackManager connectCallbackManager;
 
     @Before
     public void setUp() {
-        connectables = new Register<>();
         connectCallbackManager = new ConnectCallbackManager(connectables);
     }