--- a/core/src/test/java/com/novoda/merlin/ConnectivityCallbacksTest.java
+++ b/core/src/test/java/com/novoda/merlin/ConnectivityCallbacksTest.java
@@ -8,22 +8,14 @@ import android.os.Build;
 import android.support.annotation.RequiresApi;
 
 import org.junit.Before;
-import org.junit.Rule;
 import org.junit.Test;
-import org.mockito.ArgumentCaptor;
 import org.mockito.InOrder;
-import org.mockito.Mock;
-import org.mockito.junit.MockitoJUnit;
-import org.mockito.junit.MockitoRule;
 
-import static com.google.common.truth.Truth.assertThat;
 import static org.mockito.BDDMockito.given;
 import static org.mockito.Mockito.*;
 
 @RequiresApi(api = Build.VERSION_CODES.LOLLIPOP)
 public class ConnectivityCallbacksTest {
-
-    private static final ConnectivityChangeEvent ANY_CONNECTIVITY_CHANGE_EVENT = ConnectivityChangeEvent.createWithoutConnection();
 
     private static final boolean CONNECTED = true;
     private static final boolean DISCONNECTED = false;
@@ -37,53 +29,71 @@ public class ConnectivityCallbacksTest {
     private static final boolean CAN_NOTIFY = true;
     private static final boolean CANNOT_NOTIFY = false;
 
-    @Rule
-    public MockitoRule mockitoRule = MockitoJUnit.rule();
-
-    @Mock
-    private ConnectivityManager connectivityManager;
-    @Mock
-    private MerlinService.ConnectivityChangesNotifier connectivityChangesNotifier;
-    @Mock
-    private Network network;
-    @Mock
-    private ConnectivityChangeEventExtractor extractor;
+    private final ConnectivityManager connectivityManager = mock(ConnectivityManager.class);
+    private final MerlinService.ConnectivityChangesNotifier connectivityChangesNotifier = mock(MerlinService.ConnectivityChangesNotifier.class);
+    private final Network network = mock(Network.class);
 
     private ConnectivityCallbacks networkCallbacks;
 
     @Before
     public void setUp() {
         given(connectivityChangesNotifier.canNotify()).willReturn(CAN_NOTIFY);
-        given(extractor.extractFrom(any(Network.class))).willReturn(ANY_CONNECTIVITY_CHANGE_EVENT);
-
+        ConnectivityChangeEventExtractor extractor = new ConnectivityChangeEventExtractor(connectivityManager);
         networkCallbacks = new ConnectivityCallbacks(connectivityChangesNotifier, extractor);
     }
 
     @Test
     public void givenConnectedNetworkInfo_whenNetworkIsAvailable_thenNotifiesOfConnectedNetwork() {
-        NetworkInfo connectedNetworkInfo = givenNetworkInfoWith(CONNECTED, ANY_REASON, ANY_EXTRA_INFO);
+        NetworkInfo networkInfo = givenNetworkInfoWith(CONNECTED, ANY_REASON, ANY_EXTRA_INFO);
 
         networkCallbacks.onAvailable(network);
 
-        thenNotifies();
+        verify(connectivityChangesNotifier).notify(ConnectivityChangeEvent.createWithNetworkInfoChangeEvent(
+                networkInfo.isConnected(), networkInfo.getExtraInfo(), networkInfo.getReason()
+        ));
     }
 
     @Test
     public void givenDisconnectedNetworkInfo_whenLosingNetwork_thenNotifiesOfDisconnectedNetwork() {
-        NetworkInfo disconnectedNetworkInfo = givenNetworkInfoWith(DISCONNECTED, ANY_REASON, ANY_EXTRA_INFO);
+        NetworkInfo networkInfo = givenNetworkInfoWith(DISCONNECTED, ANY_REASON, ANY_EXTRA_INFO);
 
         networkCallbacks.onLosing(network, MAX_MS_TO_LIVE);
 
-        thenNotifies();
+        verify(connectivityChangesNotifier).notify(ConnectivityChangeEvent.createWithNetworkInfoChangeEvent(
+                networkInfo.isConnected(), networkInfo.getExtraInfo(), networkInfo.getReason()
+        ));
     }
 
     @Test
     public void givenDisconnectedNetworkInfo_whenNetworkIsLost_thenNotifiesOfLostNetwork() {
-        NetworkInfo disconnectedNetworkInfo = givenNetworkInfoWith(DISCONNECTED, ANY_REASON, ANY_EXTRA_INFO);
+        NetworkInfo networkInfo = givenNetworkInfoWith(DISCONNECTED, ANY_REASON, ANY_EXTRA_INFO);
 
         networkCallbacks.onLost(network);
 
-        thenNotifies();
+        verify(connectivityChangesNotifier).notify(ConnectivityChangeEvent.createWithNetworkInfoChangeEvent(
+                networkInfo.isConnected(), networkInfo.getExtraInfo(), networkInfo.getReason()
+        ));
+    }
+
+    @Test
+    public void givenNoNetworkInfo_whenNetworkIsAvailable_thenNotifiesOfConnectedNetwork() {
+        networkCallbacks.onAvailable(network);
+
+        verify(connectivityChangesNotifier).notify(ConnectivityChangeEvent.createWithoutConnection());
+    }
+
+    @Test
+    public void givenNoNetworkInfo_whenLosingNetwork_thenNotifiesOfDisconnectedNetwork() {
+        networkCallbacks.onLosing(network, MAX_MS_TO_LIVE);
+
+        verify(connectivityChangesNotifier).notify(ConnectivityChangeEvent.createWithoutConnection());
+    }
+
+    @Test
+    public void givenNoNetworkInfo_whenNetworkIsLost_thenNotifiesOfLostNetwork() {
+        networkCallbacks.onLost(network);
+
+        verify(connectivityChangesNotifier).notify(ConnectivityChangeEvent.createWithoutConnection());
     }
 
     @Test
@@ -122,19 +132,12 @@ public class ConnectivityCallbacksTest {
     @TargetApi(Build.VERSION_CODES.LOLLIPOP)
     private NetworkInfo givenNetworkInfoWith(boolean connected, String reason, String extraInfo) {
         NetworkInfo networkInfo = mock(NetworkInfo.class);
-
         given(networkInfo.isConnected()).willReturn(connected);
         given(networkInfo.getReason()).willReturn(reason);
         given(networkInfo.getExtraInfo()).willReturn(extraInfo);
         given(connectivityManager.getNetworkInfo(network)).willReturn(networkInfo);
 
         return networkInfo;
-    }
-
-    private void thenNotifies() {
-        ArgumentCaptor<ConnectivityChangeEvent> argumentCaptor = ArgumentCaptor.forClass(ConnectivityChangeEvent.class);
-        verify(connectivityChangesNotifier).notify(argumentCaptor.capture());
-        assertThat(argumentCaptor.getValue()).isEqualTo(ANY_CONNECTIVITY_CHANGE_EVENT);
     }
 
 }