--- a/core/src/test/java/com/novoda/merlin/ConnectivityChangeEventCreatorTest.java
+++ b/core/src/test/java/com/novoda/merlin/ConnectivityChangeEventCreatorTest.java
@@ -10,7 +10,7 @@ import org.mockito.Mock;
 import org.mockito.junit.MockitoJUnit;
 import org.mockito.junit.MockitoRule;
 
-import static org.fest.assertions.api.Assertions.assertThat;
+import static com.google.common.truth.Truth.assertThat;
 import static org.mockito.BDDMockito.given;
 
 public class ConnectivityChangeEventCreatorTest {