--- a/core/src/test/java/com/novoda/merlin/EndpointPingerTest.java
+++ b/core/src/test/java/com/novoda/merlin/EndpointPingerTest.java
@@ -1,28 +1,19 @@
 package com.novoda.merlin;
 
 import org.junit.Before;
-import org.junit.Rule;
 import org.junit.Test;
-import org.mockito.Mock;
-import org.mockito.junit.MockitoJUnit;
-import org.mockito.junit.MockitoRule;
 
 import static org.mockito.BDDMockito.given;
+import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.verify;
 
 public class EndpointPingerTest {
 
     private static final Endpoint ENDPOINT = Endpoint.from("any endpoint");
 
-    @Rule
-    public MockitoRule mockitoRule = MockitoJUnit.rule();
-
-    @Mock
-    private PingTask pingTask;
-    @Mock
-    private PingTaskFactory pingTaskFactory;
-    @Mock
-    private EndpointPinger.PingerCallback pingerCallback;
+    private final PingTask pingTask = mock(PingTask.class);
+    private final PingTaskFactory pingTaskFactory = mock(PingTaskFactory.class);
+    private final EndpointPinger.PingerCallback pingerCallback = mock(EndpointPinger.PingerCallback.class);
 
     private EndpointPinger endpointPinger;