--- a/core/src/test/java/com/novoda/merlin/MerlinServiceDependencyMissingExceptionMatcher.java
+++ b/core/src/test/java/com/novoda/merlin/MerlinServiceDependencyMissingExceptionMatcher.java
@@ -8,7 +8,7 @@ class MerlinServiceDependencyMissingExceptionMatcher extends BaseMatcher<MerlinS
     private String expectedMessage;
     private Throwable expectedCause;
 
-    public static MerlinServiceDependencyMissingExceptionMatcher from(Class expectedDependency) {
+    static MerlinServiceDependencyMissingExceptionMatcher from(Class expectedDependency) {
         MerlinServiceDependencyMissingException expectedException = MerlinServiceDependencyMissingException.missing(expectedDependency);
         return new MerlinServiceDependencyMissingExceptionMatcher(expectedException);
     }