--- a/core/src/test/java/com/novoda/merlin/MerlinsBeardTest.java
+++ b/core/src/test/java/com/novoda/merlin/MerlinsBeardTest.java
@@ -14,7 +14,7 @@ import org.mockito.Mockito;
 import org.mockito.junit.MockitoJUnit;
 import org.mockito.junit.MockitoRule;
 
-import static org.fest.assertions.api.Assertions.assertThat;
+import static com.google.common.truth.Truth.assertThat;
 import static org.mockito.BDDMockito.given;
 
 public class MerlinsBeardTest {