--- a/core/src/test/java/com/novoda/merlin/NetworkStatusRetrieverTest.java
+++ b/core/src/test/java/com/novoda/merlin/NetworkStatusRetrieverTest.java
@@ -7,7 +7,7 @@ import org.mockito.Mock;
 import org.mockito.junit.MockitoJUnit;
 import org.mockito.junit.MockitoRule;
 
-import static org.fest.assertions.api.Assertions.assertThat;
+import static com.google.common.truth.Truth.assertThat;
 import static org.mockito.BDDMockito.given;
 import static org.mockito.Mockito.verify;