--- a/dependencies.gradle
+++ b/dependencies.gradle
@@ -21,7 +21,7 @@ ext {
             test : [
                     jUnit  : 'junit:junit:4.12',
                     mockito: 'org.mockito:mockito-core:2.15.0',
-                    fest   : 'org.easytesting:fest-assert-core:2.0M8'
+                    truth  : 'com.google.truth:truth:0.39'
             ]
     ]
 }