… core/src/test/java/com/novoda/merlin/NetworkStatusTest.java
… core/src/test/java/com/novoda/merlin/PingTest.java
… core/src/test/java/com/novoda/merlin/RegistrarTest.java
… core/src/test/java/com/novoda/merlin/ResponseCodeValidatorTest.java
… merlin-rxjava/src/test/java/com/novoda/merlin/MerlinObservableTest.java
… merlin-rxjava2/src/test/java/com/novoda/merlin/MerlinFlowableTest.java