## Problem
Fest is outdated and not actively supported anymore 😬 

## Solution
Use Truth assertions instead. https://github.com/google/truth

Remove Mockito Annotation from codebase, it's slower than just specifying when adding fields due to the annotation processing. 

## Screen Capture
Just running the core batch of tests between master and this branch.

Before | After 
--- | ---
![screen shot 2018-02-14 at 16 11 15](https://user-images.githubusercontent.com/3380092/36214582-beaca32c-11a1-11e8-8ed7-b1e3aede6811.png) | ![screen shot 2018-02-14 at 16 10 46](https://user-images.githubusercontent.com/3380092/36214583-beca7b9a-11a1-11e8-9fe3-1e515b0601dc.png)


### Paired with
Nobody.