@Mecharyry 2018-02-14
@rock3r we started doing this on an internal project, I'm sure I can find the numbers that we used on that project. On this project it is negligible because of the number of tests. I prefer this because now we have private final fields, so enforcing some of our coding standards on tests, rather than just forgetting about them.