@Hutch4 2018-02-15
" I'd prioritise legibility."

I don't think the Mockito annotations add more readability. 
So is all this perf talk even needed? 

I say merge it. and have a deeper discussion of how to properly execute a perf analysis later.