@rock3r 2018-02-15
I've chatted with @Mecharyry and he prefers having final fields. The numbers only conclusively proved that this doesn’t make anything considerably faster or slower despite the claims of the article that started it all. Besides, legibility is not _that_ much better or worse either way and it'd boil down to personal preferences, so I respect his preference. **SHIPIT** :shipit: