@tasomaniac 2018-02-16
@rock3r amazing detailed investigation. I personally think that it does not worth that much investigating. It's all personal preferences, really. But I would be curious doing this on a bigger project where we have thousands of tests and the duration is minutes not seconds. 

Also regarding style, this comes very handy with Kotlin. Because otherwise you would have `lateinit var`. Not only that is ugly, but also every usage of such mocked instance is underlined (because it is a var)